requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.25
aiohttp==3.9.1
Jinja2==3.1.2
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3 